4.21 done

+++ part4/controllers/Blogs.js
@@ -49,7 +49,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(request.params.id);
+
+    console.log(user.id.toString(), blog.user.toString())
+    if(user.id.toString() === blog.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id);
+    } else {
+        response.status(401).json({ error : "invalid token" })
+    }
     response.status(204).end();
 })
